# medical-project-backend/backend/app/static/openapi.yaml
openapi: 3.1.0
info:
  title: Patient Management API
  version: "1.2.0"
  description: >
    API for managing patients, including creating new patients, updating existing patient records (with version history), medical history, medical visits, and retrieving detailed patient information.
servers:
  - url: https://8dj6t53su8wv.share.zrok.io
    description: Local development server
...
paths:
...
  /api/patients/{medical_insurance_id}/visits:
    post:
      security:
        - BearerAuth: []
      tags:
        - Medical Visits
      summary: Add Medical Visit Record
      description: >
        Adds a new medical visit record for a patient identified by their medical_insurance_id.
        This operation requires authentication and appropriate permissions.
      parameters:
        - in: path
          name: medical_insurance_id
          required: true
          schema:
            type: string
            example: "INS123456"
          description: The medical insurance ID of the patient.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisitCreate'
            examples:
              Example1:
                summary: Sample Medical Visit Creation
                value:
                  establishment_id: "e1234567-89ab-cdef-0123-456789abcdef"
                  doctor_id: "2d3cfc26-4958-4723-acf8-9799502c4d7d"
                  visit_date: "2023-06-20"
                  diagnostic: "Common cold with mild fever"
                  treatment: "Rest, hydration, and over-the-counter pain relievers"
                  summary: "Patient presented with cold symptoms. Recommended home care and rest."
                  notes: "Follow-up in one week if symptoms persist"
      responses:
        '201':
          description: Medical visit record created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitCreateResponse'
              examples:
                SuccessResponse:
                  summary: Successful Creation
                  value:
                    visit_id: "194400ed-b90d-4735-9fa2-73443c2e5700"
        '400':
          description: Bad Request - Validation Error or Foreign Key Violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/patients/{medical_insurance_id}/visits/{visit_id}:
    put:
      security:
        - BearerAuth: []
      tags:
        - Medical Visits
      summary: Update Medical Visit Record
      description: >
        Updates an existing medical visit record for a patient while preserving the previous version.
        This operation requires authentication and appropriate permissions.
      parameters:
        - in: path
          name: medical_insurance_id
          required: true
          schema:
            type: string
            example: "INS123456"
          description: The medical insurance ID of the patient.
        - in: path
          name: visit_id
          required: true
          schema:
            type: string
            format: uuid
            example: "194400ed-b90d-4735-9fa2-73443c2e5700"
          description: The unique identifier of the medical visit record.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisitUpdate'
            examples:
              Example1:
                summary: Sample Medical Visit Update
                value:
                  establishment_id: "e1234567-89ab-cdef-0123-456789abcdef"
                  doctor_id: "2d3cfc26-4958-4723-acf8-9799502c4d7d"
                  visit_date: "2023-06-20"
                  diagnostic: "Upper respiratory infection"
                  treatment: "Prescribed antibiotics and continued rest"
                  summary: "Follow-up visit for cold symptoms. Condition worsened to respiratory infection."
                  notes: "Complete full course of antibiotics. Return if fever persists."
      responses:
        '201':
          description: Medical visit record updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitUpdateResponse'
              examples:
                SuccessResponse:
                  summary: Successful Update
                  value:
                    visit_id: "194400ed-b90d-4735-9fa2-73443c2e5700"
                    patient_id: "INS123456"
                    establishment_id: "e1234567-89ab-cdef-0123-456789abcdef"
                    doctor_id: "2d3cfc26-4958-4723-acf8-9799502c4d7d"
                    visit_date: "2023-06-20"
                    diagnostic: "Upper respiratory infection"
                    treatment: "Prescribed antibiotics and continued rest"
                    summary: "Follow-up visit for cold symptoms. Condition worsened to respiratory infection."
                    notes: "Complete full course of antibiotics. Return if fever persists."
                    created_at: "2023-06-20T10:30:00Z"
        '400':
          description: Bad Request - Validation Error or Foreign Key Violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found - Visit record does not exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/visits/{visit_id}:
    delete:
      security:
        - BearerAuth: []
      tags:
        - Medical Visits
      summary: Hide a Medical Visit Record
      description: >
        Makes a medical visit record hidden in the system without actually deleting it.
        This preserves data integrity while removing it from standard queries.
      parameters:
        - in: path
          name: visit_id
          required: true
          schema:
            type: string
            format: uuid
          description: The unique identifier of the medical visit record to hide.
      responses:
        '200':
          description: Medical visit record successfully hidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
              examples:
                SuccessResponse:
                  summary: Successful Hiding
                  value:
                    status: "success"
        '403':
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found - Medical visit record does not exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      security:
        - BearerAuth: []
      tags:
        - Medical Visits
      summary: Update Medical Visit Record (Alternative)
      description: >
        Updates an existing medical visit record without needing the medical_insurance_id.
        This is an alternative to the /patients/{medical_insurance_id}/visits/{visit_id} endpoint.
      parameters:
        - in: path
          name: visit_id
          required: true
          schema:
            type: string
            format: uuid
            example: "194400ed-b90d-4735-9fa2-73443c2e5700"
          description: The unique identifier of the medical visit record.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisitUpdate'
      responses:
        '201':
          description: Medical visit record updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitUpdateResponse'
        '400':
          description: Bad Request - Validation Error or Foreign Key Violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not Found - Visit record does not exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/auth/login
  schemas:
...
    VisitCreate:
      type: object
      description: Schema for creating a new medical visit record.
      properties:
        establishment_id:
          type: string
          format: uuid
          example: "e1234567-89ab-cdef-0123-456789abcdef"
        doctor_id:
          type: string
          format: uuid
          example: "2d3cfc26-4958-4723-acf8-9799502c4d7d"
        visit_date:
          type: string
          format: date
          example: "2022-02-20"
        diagnostic:
          type: string
          example: "Common cold with mild fever"
          nullable: true
        treatment:
          type: string
          example: "Rest, hydration, and over-the-counter pain relievers"
          nullable: true
        summary:
          type: string
          example: "Patient presented with cold symptoms. Recommended home care and rest."
        notes:
          type: string
          example: "Follow-up in one week if symptoms persist"
          nullable: true
      required:
        - establishment_id
        - doctor_id
        - visit_date
        - summary
      additionalProperties: false
    VisitUpdate:
      type: object
      description: Schema for updating an existing medical visit record.
      properties:
        establishment_id:
          type: string
          format: uuid
          example: "e1234567-89ab-cdef-0123-456789abcdef"
        doctor_id:
          type: string
          format: uuid
          example: "2d3cfc26-4958-4723-acf8-9799502c4d7d"
        visit_date:
          type: string
          format: date
          example: "2022-02-20"
        diagnostic:
          type: string
          example: "Common cold with mild fever"
          nullable: true
        treatment:
          type: string
          example: "Rest, hydration, and over-the-counter pain relievers"
          nullable: true
        summary:
          type: string
          example: "Patient presented with cold symptoms. Recommended home care and rest."
        notes:
          type: string
          example: "Follow-up in one week if symptoms persist"
          nullable: true
      required:
        - establishment_id
        - doctor_id
        - visit_date
        - summary
      additionalProperties: false
    VisitCreateResponse:
      type: object
      description: Response schema for a newly created medical visit record.
      properties:
        visit_id:
          type: string
          format: uuid
          example: "194400ed-b90d-4735-9fa2-73443c2e5700"
      required:
        - visit_id
      additionalProperties: false
    VisitUpdateResponse:
      type: object
      description: Response schema for an updated medical visit record.
      properties:
        visit_id:
          type: string
          format: uuid
          example: "194400ed-b90d-4735-9fa2-73443c2e5700"
        patient_id:
          type: string
          example: "INS123456"
        establishment_id:
          type: string
          format: uuid
          example: "e1234567-89ab-cdef-0123-456789abcdef"
        doctor_id:
          type: string
          format: uuid
          example: "2d3cfc26-4958-4723-acf8-9799502c4d7d"
        visit_date:
          type: string
          format: date
          example: "2022-02-20"
        diagnostic:
          type: string
          example: "Common cold with mild fever"
          nullable: true
        treatment:
          type: string
          example: "Rest, hydration, and over-the-counter pain relievers"
          nullable: true
        summary:
          type: string
          example: "Patient presented with cold symptoms. Recommended home care and rest."
        notes:
          type: string
          example: "Follow-up in one week if symptoms persist"
          nullable: true
        created_at:
          type: string
          format: date-time
          example: "2022-02-20T10:30:00Z"
      required:
        - visit_id
        - patient_id
        - establishment_id
        - doctor_id
        - visit_date
        - summary
      additionalProperties: false
...
    ErrorResponse:
      type: object
      description: Schema for error responses.
      properties:
        error:
          type: string
          example: "Invalid foreign key reference."
      required:
        - error
      additionalProperties: false
...
    StatusResponse:
      type: object
      description: Simple status response.
      properties:
        status:
          type: string
          example: "success"
      required:
        - status
      additionalProperties: false
...